Advance H.264/AV1 decode paths and AAC codebook integration
- H.264: correct 8x8 CAVLC luma residual zigzag interleaving, add PPS
  parse diagnostics, suppress noisy debug output, and allow parser-style
  clippy lints crate-wide.
- AV1: fix force_integer_mv semantics per spec 6.8.2 and align frame
  header bitstream test with keyframe sizing/loop-restoration layout.
- AAC: wire generated codebooks module into the public API.
- Realtime/volumetric: minor clippy and test cleanup.

+++ tpt-kinetix-h264/src/entropy.rs
@@ -485,7 +485,7 @@ impl Intra4x4PredModeCabacContext {
         }
         let mut mode = 0u8;
         for i in 0..3u8 {
-            mode += (dec.decode_decision(&mut self.rem_ctx) as u8) << i;
+            mode += dec.decode_decision(&mut self.rem_ctx) << i;
         }
         mode + (mode >= pred_mode) as u8
     }
